Enhancing REST API, so its more secure

+++ src/main/java/com/example/zahaawiiblog/commentsFeature/Controller/CommentsController.java
@@ -4,13 +4,18 @@ package com.example.zahaawiiblog.commentsFeature.Controller;
 import com.example.zahaawiiblog.commentsFeature.DTO.CommentsDTO;
 import com.example.zahaawiiblog.commentsFeature.entity.Comments;
 import com.example.zahaawiiblog.commentsFeature.service.CommentsService;
+import com.example.zahaawiiblog.logginFeature.service.LoggingService;
+import com.example.zahaawiiblog.securityFeature.Entity.UserInfo;
 import com.example.zahaawiiblog.service.BlogService;
 import com.example.zahaawiiblog.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/v1/comments")
@@ -18,18 +23,25 @@ import java.util.List;
 public class CommentsController {
 
     private final CommentsService commentsService;
-    private final BlogService blogService;
     private final UserService userService;
+    private final LoggingService loggingService;
 
-    public CommentsController (CommentsService commentsService, BlogService blogService, UserService userService) {
+    public CommentsController (CommentsService commentsService, UserService userService, LoggingService loggingService) {
         this.commentsService = commentsService;
-        this.blogService = blogService;
         this.userService = userService;
+        this.loggingService = loggingService;
     }
 
     @PostMapping("/addcomment")
-    public ResponseEntity<CommentsDTO> addComment (@RequestBody CommentsDTO comment) {
+    public ResponseEntity<?> addComment (@RequestBody CommentsDTO comment) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        UserInfo user = userService.findUserByUsername(auth.getName()).orElseThrow(() -> new RuntimeException("User not found"));
+        if(auth == null || !auth.isAuthenticated()) {
+            loggingService.log(0L, " an unathorized user tried to add a comment", "unathorized user", 4L);
+            return new ResponseEntity<>("User not logged in", HttpStatus.UNAUTHORIZED);
+        }
         commentsService.add(comment);
+        loggingService.log(user.getUserId(), user.getName() + ": deleted a comment with id: " + comment.commentId() , user.getName(), 3L);
         return new ResponseEntity<>(comment,HttpStatus.OK);
     }
 
@@ -41,7 +53,14 @@ public class CommentsController {
 
     @DeleteMapping("/delete/{commentid}")
     public ResponseEntity<?> deleteComment(@PathVariable Long commentid) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        UserInfo user = userService.findUserByUsername(auth.getName()).orElseThrow(() -> new RuntimeException("User not found"));
+        if(auth == null || !auth.isAuthenticated()) {
+            loggingService.log(user.getUserId(),"an unathorized user: tried to deleted a comment with id: " + commentid , "unathorized user", 4L);
+            return new ResponseEntity<>("User not logged in", HttpStatus.UNAUTHORIZED);
+        }
         commentsService.deleteByCommentId(commentid);
+        loggingService.log(user.getUserId(), user.getName() + ": deleted a comment with id: " + commentid , user.getName(), 3L);
         return new ResponseEntity<>("Comment deleted", HttpStatus.OK);
     }
